Rename setlocation to setgeopoint
https://forum.opendatakit.org/t/spec-addition-proposal-location-preload/10768/20
