+++ collect_app/src/androidTest/java/org/odk/collect/android/formentry/backgroundlocation/SetGeopointActionTest.java
@@ -25,11 +25,11 @@ import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static androidx.test.espresso.matcher.ViewMatchers.withId;
 import static androidx.test.espresso.matcher.ViewMatchers.withText;
 
-public class SetLocationActionTest {
-    private static final String SETLOCATION_ACTION_FORM = "setlocation-action.xml";
+public class SetGeopointActionTest {
+    private static final String SETGEOPOINT_ACTION_FORM = "setgeopoint-action.xml";
 
     @Rule
-    public IntentsTestRule<FormEntryActivity> activityTestRule = FormLoadingUtils.getFormActivityTestRuleFor(SETLOCATION_ACTION_FORM);
+    public IntentsTestRule<FormEntryActivity> activityTestRule = FormLoadingUtils.getFormActivityTestRuleFor(SETGEOPOINT_ACTION_FORM);
 
     @Rule
     public RuleChain copyFormChain = RuleChain
@@ -39,7 +39,7 @@ public class SetLocationActionTest {
                     Manifest.permission.ACCESS_FINE_LOCATION)
             )
             .around(new ResetStateRule())
-            .around(new CopyFormRule(SETLOCATION_ACTION_FORM, "forms"));
+            .around(new CopyFormRule(SETGEOPOINT_ACTION_FORM, "forms"));
 
     @Before
     @After
